* New /set, /SET SWITCH_CHANNELS_BETWEEN_WINDOWS * C language issues (const correctness, integer size correctness) * Revamp channel structure to hold window refnum and not window pointer * Add "current channel counter" to channel structure. * Determine "current channel" by election. * Garbage collect a lot of now unnecessary code. * Add get_echannel_by_refnum(), returns window's "elected" current channel * Add internal function is_waiting_for_channel() * Add internal function move_waiting_channel() * Add internal function get_winref_by_bound_channel() * Add internal function get_bound_channel_by_refnum() * Convert uses of get_channel_by_refnum() to get_echannel_by_refnum() * Fix $winbound() to work when given a window refnum * Fix valgrind issue with $regcomp() * Remove sanity checks from traverse_all_channels(); handled elsewhere now * Change traverse_all_channels() with "this server" or "not this server" arg * Nuke (Window *)->current_channel.  Handled by channels now.  HUZZAH! * Implement /ON SWITCH_WINDOWS again.

+++ include/ircaux.h
@@ -7,7 +7,7 @@
  *
  * See the COPYRIGHT file, or do a HELP IRCII COPYRIGHT 
  *
- * @(#)$Id: ircaux.h,v 1.41 2003/03/24 09:20:29 jnelson Exp $
+ * @(#)$Id: ircaux.h,v 1.42 2003/03/29 08:10:22 jnelson Exp $
  */
 
 #ifndef _IRCAUX_H_
@@ -157,7 +157,7 @@ int	charcount		(const char *, char);
 void	beep_em			(int);
 void	strip_control		(const char *, char *);
 const char *strfill		(char, int);
-char *	encode			(const char *, int);
+char *	encode			(const char *, size_t);
 char *	decode			(const char *);
 char *	chomp			(char *);
 int 	opento			(const char *, int, off_t);
